Split table component into multiple components and add search bar
