+++ app/package.json
@@ -1,11 +1,11 @@
 {
   "name": "csx",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@chakra-ui/react": "^1.6.1",
-    "@emotion/react": "^11.1.5",
-    "@emotion/styled": "^11.1.5",
+    "@chakra-ui/react": "^1.8.8",
+    "@emotion/react": "^11.9.3",
+    "@emotion/styled": "^11.9.3",
     "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
@@ -21,7 +21,7 @@
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^3.4.0",
     "formik": "^2.2.9",
-    "framer-motion": "^4.0.3",
+    "framer-motion": "^4.1.17",
     "html2canvas": "*",
     "husky": "^6.0.0",
     "lint-staged": "^11.0.0",
